refactor: 토큰 처리 공통화
